Revert "Merge branch 'Denizs-Branch' of https://github.com/metehanCekci/HexandVexDaSteamEdition into Denizs-Branch"
This reverts commit ba3cf076f9ac18a51a05908d5943a8c9b3776bce, reversing
changes made to 758e679d13ea952cfee578fbc7642c30b41e39b2.

+++ Assets/Scripts/SpawnerBossAi.cs
@@ -31,7 +31,7 @@ public class SpawnerBossAI : MonoBehaviour
     
     private bool isTransitioning = false; 
     private bool isSummoning = false; 
-    private long previousHP;
+    private int previousHP; 
 
     private Tilemap groundMap;
     private Tilemap bossWarningMap;
